+++ Source/EventFlow/Core/ITransientFaultHandler.cs
@@ -21,25 +21,19 @@
 // CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 
 using System;
+using System.Threading;
+using System.Threading.Tasks;
 
 namespace EventFlow.Core
 {
-    public class Retry
+    public interface ITransientFaultHandler
     {
-        public static Retry Yes { get { return new Retry(true, TimeSpan.Zero); } }
-        public static Retry YesAfter(TimeSpan retryAfter) { return new Retry(true, retryAfter); }
-        public static Retry No { get { return new Retry(false, TimeSpan.Zero); } }
+        void Use<TRetryStrategy>(Action<TRetryStrategy> configureStrategy = null)
+            where TRetryStrategy : IRetryStrategy;
 
-        public bool ShouldBeRetried { get; set; }
-        public TimeSpan RetryAfter { get; set; }
-
-        private Retry(bool shouldBeRetried, TimeSpan retryAfter)
-        {
-            if (retryAfter != TimeSpan.Zero && retryAfter != retryAfter.Duration()) throw new ArgumentOutOfRangeException("retryAfter");
-            if (!shouldBeRetried && retryAfter != TimeSpan.Zero) throw new ArgumentException("Invalid combination");
-
-            ShouldBeRetried = shouldBeRetried;
-            RetryAfter = retryAfter;
-        }
+        Task<T> TryAsync<T>(
+            Func<CancellationToken, Task<T>> action,
+            Label label,
+            CancellationToken cancellationToken);
     }
 }