+++ RELEASE_NOTES.md
@@ -2,6 +2,9 @@
 
 * Breaking: `ValueObject` now uses public properties instead of both
   private and public fields
+* Breaking: MSSQL transient exceptions are now retried
+* New: `ITransientFaultHandler` added along with default retry strategies
+  for optimistic concurrency and MSSQL transient exceptions
 * New: Release notes added to NuGet packages
 * New: Better logging and more descriptive exceptions
 * Fixed: Unchecked missing in `ValueObject` when claculating hash